lottery: mnist mlp plotting
Finished refactoring the training and plotting to be entirely separate. Although
this is not enitrely necessary for MNIST, it will help to shorten the
development cycle on models/datasets that are more expensive to train,
eg. CIFAR and MNIST. This change notably makes a storage/compute
tradeoff that makes it more onerous to create nice videos, but this is
something that we'll have to figure out in the future, and may only be
feasible with MNIST anyhow.